rename: 캘린더 컴포넌트 및 타입 위치 변경-#27
